refactored the Encounter component out of App.js to enable following along with https://appdividend.com/2017/07/22/react-firebase-tutorial to complete the firebase database integration - this closes #30
