+++ timeline/src/App.js
@@ -1,10 +1,10 @@
 import React, { Component } from 'react';
 import './App.css';
-
+import Encounter from './components/Encounter';
 class App extends Component {
   constructor() {
     super();
-    // TODO: pull in data - local at first, remotely once the data binding logic settles
+    // TODO: pull in remote data
     this.state = {
       encounters: [
         {
@@ -127,35 +127,3 @@ class App extends Component {
 }
 
 export default App;
-
-// This is the repeating element that displays one or more objects in the Timeline, one per encounter
-class Encounter extends Component {
-  render() {
-    return (
-      <div className="Encounter">
-        <div className="Person-image">
-          {/* TODO: insert appropriately parameterized <img> tag */}
-        </div>
-        <div className="Encounter_data">
-          <div className="Encounter-first-row">
-            <div className="Person-name">
-              Who: {this.props.valuePerson}
-            </div>
-            <div className="Encounter-date">
-              When: {this.props.valueDate}
-            </div>
-          </div>
-          <div className="Encounter-second-row">
-            <div className="Encounter-event">
-              Where: {this.props.valueEvent}
-            </div>
-          </div>
-        </div>
-        {/* <div className="Encounter-navigation">
-          <button>></button>
-        </div> */}
-        <hr />
-      </div>
-    );
-  }
-}
